pyproject.toml
[tool.poetry]
name = "nexus-ai-sdk"
version = "0.1.0"
description = "Python SDK for Nexus AI Platform - A unified AI capabilities platform"
authors = ["Nexus AI Team <support@nexus-ai.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://nexus-ai.juncai-ai.com"
repository = "https://github.com/nexus-ai/python-sdk"
keywords = ["ai", "api", "sdk", "nexus", "openai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
packages = [{include = "nexusai"}]

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.25.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
typing-extensions = "^4.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-httpx = "^0.24.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
mypy = "^1.7.0"
ruff = "^0.1.0"
types-requests = "^2.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=nexusai",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["nexusai"]
omit = ["tests/*", "nexusai/__version__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
